second round of addressing PR comments
