+++ react_frontend/src/pages/games/memoryVault/imports/audioFiles.js
@@ -1,0 +1,26 @@
+import alarm from '../audio/alarm.mp3';
+import block from '../audio/block.mp3';
+import boat from '../audio/boat.mp3';
+import brush from '../audio/brush.mp3';
+import cloud from '../audio/cloud.mp3';
+import door from '../audio/door.mp3';
+import egg from '../audio/egg.mp3';
+import frame from '../audio/frame.mp3';
+import globe from '../audio/globe.mp3';
+import grass from '../audio/grass.mp3';
+import hammock from '../audio/hammock.mp3';
+import kettle from '../audio/kettle.mp3';
+import leaf from '../audio/leaf.mp3';
+import lemon from '../audio/lemon.mp3';
+import mouse from '../audio/mouse.mp3';
+import notebook from '../audio/notebook.mp3';
+import pineapple from '../audio/pineapple.mp3';
+import rainbow from '../audio/rainbow.mp3';
+import ruler from '../audio/ruler.mp3';
+import salt from '../audio/salt.mp3';
+import shark from '../audio/shark.mp3';
+
+export {
+  alarm, block, boat, brush, cloud, door, egg, frame, globe, grass, hammock,
+  kettle, leaf, lemon, mouse, notebook, pineapple, rainbow, ruler, salt, shark
+};
