+++ react_frontend/src/pages/games/memoryVault/imports/imageFiles.js
@@ -1,0 +1,26 @@
+import apple from '../pictures/apple.png';
+import bell from '../pictures/bell.png';
+import bowl from '../pictures/bowl.png';
+import button from '../pictures/button.png';
+import candle from '../pictures/candle.png';
+import carrot from '../pictures/carrot.png';
+import chair from '../pictures/chair.png';
+import fan from '../pictures/fan.png';
+import hammer from '../pictures/hammer.png';
+import key from '../pictures/key.png';
+import kite from '../pictures/kite.png';
+import knife from '../pictures/knife.png';
+import ladder from '../pictures/ladder.png';
+import lamp from '../pictures/lamp.png';
+import map from '../pictures/map.png';
+import moon from '../pictures/moon.png';
+import pencil from '../pictures/pencil.png';
+import sock from '../pictures/sock.png';
+import star from '../pictures/star.png';
+import vase from '../pictures/vase.png';
+import volcano from '../pictures/volcano.png';
+
+export {
+  apple, bell, bowl, button, candle, carrot, chair, fan, hammer, key, kite,
+  knife, ladder, lamp, map, moon, pencil, sock, star, vase, volcano
+};
