+++ react_frontend/src/pages/games/memoryVault/imports/recallSessions.js
@@ -1,0 +1,26 @@
+const RECALL_SESSIONS = [
+    { word: "Spoon", audio: 'Rainbow', picture: 'Apple', wordHint: 'Used to eat soup', audioHint: 'Colours in the sky', pictureHint: 'A fruit'},
+    { word: "Bicycle", audio: 'Shark', picture: 'Pencil', wordHint: 'Two-wheeled ride', audioHint: 'Ocean predator', pictureHint: 'Used to write'},
+    { word: "Castle", audio: 'Notebook', picture: 'Volcano', wordHint: 'Royal home', audioHint: 'Used for writing', pictureHint: 'Erupts lava'},
+    { word: "Clock", audio: 'Globe', picture: 'Hammer', wordHint: 'Tells time', audioHint: 'Earth model', pictureHint: 'Drives nails'},
+    { word: "Crown", audio: 'Pinneaple', picture: 'Lamp', wordHint: 'Worn by royalty', audioHint: 'Spiky fruit', pictureHint: 'Light source'},
+    { word: "Flower", audio: 'Hammock', picture: 'Kite', wordHint: 'It blooms', audioHint: 'Hanging bed', pictureHint: 'Flies in the wind'},
+    { word: "Window", audio: 'Grass', picture: 'Moon', wordHint: 'Glass opening', audioHint: 'Green ground', pictureHint: 'Night light'},
+    { word: "Fire", audio: 'Leaf', picture: 'Key', wordHint: 'Hot and bright', audioHint: 'Green plant part', pictureHint: 'Used to unlock'},
+    { word: "Cookie", audio: 'Door', picture: 'Knife', wordHint: 'Sweet treat', audioHint: 'Entrance', pictureHint: 'Cutting tool'},
+    { word: "Rain", audio: 'Egg', picture: 'Star', wordHint: 'Water from sky', audioHint: 'Oval protein source', pictureHint: 'Shines at night'},
+    { word: "Locket", audio: 'Cloud', picture: 'ladder', wordHint: 'Pendant with a photo', audioHint: 'Fluffy in the sky', pictureHint: 'For climbing'},
+    { word: "Bottle", audio: 'Ruler', picture: 'Button', wordHint: 'Holds liquid', audioHint: 'Measures length', pictureHint: 'Fastens clothes'},
+    { word: "Toothbrush", audio: 'Kettle', picture: 'Candle', wordHint: 'Cleans teeth', audioHint: 'Boils water', pictureHint: 'Wax light'},
+    { word: "Drum", audio: 'Brush', picture: 'Carrot', wordHint: 'Beaten instrument', audioHint: 'Cleans or paints', pictureHint: 'Orange veggie'},
+    { word: "Rope", audio: 'Lemon', picture: 'Bell', wordHint: 'Strong cord', audioHint: 'Sour fruit', pictureHint: 'It rings'},
+    { word: "Belt", audio: 'Alarm', picture: 'Chair', wordHint: 'Waistwear', audioHint: 'Wakes you up', pictureHint: 'Used to sit on'},
+    { word: "Wheel", audio: 'Block', picture: 'Vase', wordHint: 'It rolls', audioHint: 'Solid piece', pictureHint: 'Footwear'},
+    { word: "Bridge", audio: 'Mouse', picture: 'Sock', wordHint: 'Spans gaps', audioHint: 'Small rodent', pictureHint: 'Used to drink'},
+    { word: "Honey", audio: 'Frame', picture: 'Fan', wordHint: 'Sweet from bees', audioHint: 'Holds pictures', pictureHint: 'Blows air'},
+    { word: "Glass", audio: 'Boat', picture: 'Bowl', wordHint: 'Clear material', audioHint: 'Water transport', pictureHint: 'Holds food'},
+    { word: "Basket", audio: 'Salt', picture: 'Map', wordHint: 'Holds items', audioHint: 'Adds flavour', pictureHint: 'Shows places'},
+  ];
+
+  export default RECALL_SESSIONS;
+
