+++ react_frontend/src/pages/games/memoryVault/imports/startSessions.js
@@ -1,0 +1,29 @@
+import { apple, bell, bowl, button, candle, carrot, chair, fan, hammer, key, kite, knife,
+    ladder, lamp, map, moon, pencil, sock, star, vase, volcano } from './imageFiles';
+  
+  const SESSIONS = [
+    { word: "spoon", audio: 'rainbow', picture: 'apple', pictureRef: apple },
+    { word: "bicycle", audio: 'shark', picture: 'pencil', pictureRef: pencil },
+    { word: "castle", audio: 'notebook', picture: 'volcano', pictureRef: volcano },
+    { word: "clock", audio: 'globe', picture: 'hammer', pictureRef: hammer },
+    { word: "crown", audio: 'pineapple', picture: 'lamp', pictureRef: lamp },
+    { word: "flower", audio: 'hammock', picture: 'kite', pictureRef: kite },
+    { word: "window", audio: 'grass', picture: 'moon', pictureRef: moon },
+    { word: "fire", audio: 'leaf', picture: 'key', pictureRef: key },
+    { word: "cookie", audio: 'door', picture: 'knife', pictureRef: knife },
+    { word: "rain", audio: 'egg', picture: 'star', pictureRef: star },
+    { word: "locket", audio: 'cloud', picture: 'ladder', pictureRef: ladder },
+    { word: "bottle", audio: 'ruler', picture: 'button', pictureRef: button },
+    { word: "toothbrush", audio: 'kettle', picture: 'candle', pictureRef: candle },
+    { word: "drum", audio: 'brush', picture: 'carrot', pictureRef: carrot },
+    { word: "rope", audio: 'lemon', picture: 'bell', pictureRef: bell },
+    { word: "belt", audio: 'alarm', picture: 'chair', pictureRef: chair },
+    { word: "wheel", audio: 'block', picture: 'vase', pictureRef: vase },
+    { word: "bridge", audio: 'mouse', picture: 'sock', pictureRef: sock },
+    { word: "honey", audio: 'frame', picture: 'fan', pictureRef: fan },
+    { word: "glass", audio: 'boat', picture: 'bowl', pictureRef: bowl },
+    { word: "basket", audio: 'salt', picture: 'map', pictureRef: map },
+  ];
+  
+  export default SESSIONS;
+
