+++ react_frontend/src/pages/patient/InstructionsPage.js
@@ -1,9 +1,34 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
 import titleImage from '../../assets/title.svg';
+import { db } from "../../firebaseConfig";
+import { doc, getDoc, updateDoc } from "firebase/firestore";
 import './GeneralInstructionsPage.css';
 
 const GeneralInstructionsPage = () => {
+
+  const [userId, setUserId] = useState(null);
+  const [firstPlay, setFirstPlay] = useState(false);
+
+  useEffect(() => {
+    const storedUserId = localStorage.getItem("userId");
+    if (storedUserId) {
+      setUserId(storedUserId);
+      fetchUserData(storedUserId);
+    }
+  }, []);
+
+  const fetchUserData = async (userId) => {
+    const userRef = doc(db, "users", userId);
+    const userDoc = await getDoc(userRef);
+
+    if (userDoc.exists()) {
+      if(!userDoc.data().playCount || userDoc.data().playCount === 0){
+        setFirstPlay(true);
+      }
+    }
+  };
+
   return (
     <div className="general-instructions-container">
       <div className="title-container-instructions">
@@ -22,8 +47,16 @@ const GeneralInstructionsPage = () => {
         <p className="instructions-note">
           Make sure you're in a quiet environment, free from distractions, so you can fully focus on your task!
         </p>
-        <div className="start-button-container">
-          <Link to="/patient-home-page" className="start-button">Back to Home</Link>
+        <div className="side-by-side-buttons">
+          <div className="instructions-button-container">
+            <Link to="/patient-home-page" className="start-button">Back to Home</Link>
+          </div>
+
+          {firstPlay && 
+            <div className="instructions-button-container">
+              <Link to="/lighting-calibration" className="secondary-button">Start Calibration</Link>
+            </div>
+          }
         </div>
       </div>
     </div>
