+++ react_frontend/src/pages/patient/PatientHomePage.css
@@ -143,15 +143,22 @@
 .progress-bar {
   width: 100%;
   height: 15px;
-  background-color: #d9d9d9; /* Base color of the progress bar */
+  background-color: #d9d9d9;
   border-radius: 10px;
   overflow: hidden;
   margin: 30px 0;
 }
 
+.progress-bar-value{
+  color: #f8727b;
+  font-size: 22px;
+  text-align: center;
+  margin-bottom: -20px;
+}
+
 .progress-bar-fill {
   height: 100%;
-  background-color: #f8727b; /* Filled color of the progress bar */
+  background-color: #f8727b;
   border-radius: 5px 0 0 5px;
 }
